Make each Rust tool a single file
